Added TSnackbar.setMaxWidth() method to allow overriding value
The max width value is already set through the layout but is not exposed for
clients to change it. In most cases, changing the max width is not necessary,
especially because it currently matches Google's Material guidelines. But
there may be cases where a designer would prefer to treat this as simply an
animated view that shows at the top of the screen instead of a snackbar.